Added a test for compilation with libpng before actually returning that it's ok.
Due to Apple's idea of a "useful config script", and that they placed libpng-config with hardcoded values that don't respect the SDK in a systemwide /usr/bin, this test will fail in MacOSX if not compiling for the system SDK.

i.e. "libpng-config --libs" will return -lpng15 on Lion, but that library doesn't exist in the 10.6 or 10.5 SDKs, so it breaks on linking with this SDKs in the ldflags. Similar problem with --cflags and include/png15

Anyway it's good to test compilation of a small program before deciding everything's ok.

If someone can make a test that uses --with-libpng-prefix to test for the actual headers and library without libpng-config, that'd be swell (as I would just -prefix= to the SDK). But I'll probably just statically link to a self-compiled libpng anyway if I want to compile amuleweb on MacOSX (I don't).

If you're reading this line, you have too much free time.